truck: move auth infra folder
