+++ src/app.module.ts
@@ -4,7 +4,7 @@ import * as Joi from 'joi';
 import { PrismaService } from './prisma.service';
 import { UserModule } from './modules/user/user.module';
 import { APP_GUARD } from '@nestjs/core';
-import { JwtAuthGuard } from './common/infra/guards/jwt-auth.guard';
+import { JwtAuthGuard } from './modules/auth/infra/guards/jwt-auth.guard';
 import { AuthModule } from './modules/auth/auth.module';
 
 @Module({
