+++ src/modules/auth/auth.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Post, Request, UseFilters } from '@nestjs/common';
 import { AppError } from '../../common/errors/Error';
 import { HttpExceptionFilter } from '../../common/filter/http-exception.filter';
-import { isPublic } from '../../common/infra/decorators/is-public.decorator';
+import { isPublic } from './infra/decorators/is-public.decorator';
 import { SignInService } from './services/signin.service';
 import { AuthRequest, UserToken } from './interfaces/services.interface';
